requirements.txt
Django
gunicorn
djangorestframework
djangorestframework-jwt
pymysql
flask 
requests